Use cron for autoscrim; general refactoring
