+++ infrastructure/env.Dockerfile
@@ -1,0 +1,4 @@
+FROM ubuntu:18.04
+
+ENV BC20_DB_USERNAME database_admin
+ENV BC20_DB_PASSWORD ???
