+++ infrastructure/scrimmage/crontab
@@ -1,0 +1,1 @@
+00 00,12 * * * /bin/bash /app/matchmake.sh
