+++ infrastructure/tournament.Dockerfile
@@ -1,4 +1,13 @@
-FROM bc20-matcher
+FROM bc20-env
 
-COPY app/tournament_server.py app/bracket.py app/
+# Install software dependencies
+RUN apt-get update \
+  && apt-get install -y \
+    python3 \
+    python3-pip
+RUN pip3 install --upgrade \
+    requests
+
+COPY * app/
+
 CMD ./app/tournament_server.py
