+++ infrastructure/tournament/team_names
