+++ infrastructure/worker/box/gradle.properties
